refactor: move `InputFile` to an "internal" module of `lib.rs`
The goal is to be able to reuse this file format for all targets (integration tests, benchmarks, and cli)
