+++ miden/Cargo.toml
@@ -41,9 +41,10 @@ path = "tests/integration/main.rs"
 concurrent = ["prover/concurrent", "std"]
 default = ["std"]
 executable = [
+    "std",
+    "internal",
     "dep:hex",
     "hex?/std",
-    "std",
     "dep:serde",
     "serde?/std",
     "dep:serde_derive",
@@ -55,6 +56,8 @@ executable = [
 ]
 metal = ["prover/metal", "std"]
 std = ["assembly/std", "processor/std", "prover/std", "verifier/std"]
+# For internal use, not meant to be used by users
+internal = []
 
 [dependencies]
 assembly = { package = "miden-assembly", path = "../assembly", version = "0.11", default-features = false }
