+++ miden/src/cli/prove.rs
@@ -2,11 +2,12 @@ use std::{path::PathBuf, time::Instant};
 
 use assembly::diagnostics::{IntoDiagnostic, Report, WrapErr};
 use clap::Parser;
-use miden_vm::ProvingOptions;
+use miden_vm::{internal::InputFile, ProvingOptions};
 use processor::{DefaultHost, ExecutionOptions, ExecutionOptionsError, Program};
 use stdlib::StdLibrary;
+use tracing::instrument;
 
-use super::data::{instrument, Debug, InputFile, Libraries, OutputFile, ProgramFile, ProofFile};
+use super::data::{Debug, Libraries, OutputFile, ProgramFile, ProofFile};
 
 #[derive(Debug, Clone, Parser)]
 #[clap(about = "Prove a miden program")]
